Re-enable stale news warnings for CNN
Although CNN uses what seems to have become a standard across many news
sites for specifying the publication date/time (i.e. the `content`
attributes of the `<meta>` element with the `"article:published_time"`
property), this doesn't seem to be available when the extension is
loaded.  However, CNN provides the `<meta>` element with the `pubdate`
name and the date information is stored in this element's `content`
attribute.  This change now gets the stale news warning to work again on
CNN.