Re-enable stale news warnings for the India Times
As with many other news sites, the publication date/time is stored in
the `article:published_time` meta property.  However, sometimes this
isn't seen by the stale news warning extension, hence the
`datePublished` element of the page's JSON metadata (stored in a
`<script>` element) is used as a fall-back.

The weird thing with the India Times is that when loading a page the
first time, the extension doesn't pick up any publication date/time
information, however on reload it *does*.  Odd.